feat(GCB): use Kaniko build cache
Kaniko is an tool for building containers from Dockerfiles. It caches most steps if it doesn't detect any changes. For us that means it'll caching things like the step where we download and install the `google-cloud-sdk` package. Since this step is both slow and error-prone, this will make our builds both faster and less flaky.

It does _not_ cache COPY steps that copy files from the build context. That means it won't skip over the step that copies our gradle output into the built containers.

Once it gets to an uncached step, it won't cache any steps after that (since it doesn't know how the changed input could affect those later steps). Since all our Dockerfiles have COPY has the very first step, that means essentially everything will be uncached. So we need to restructure the files to make good use of Kaniko.

The default cache time is 2 weeks. That means we won't get updates to `gcloud` (or other similar things like `apt-get` that grab some unspecified version of a package) until the cache is reset every two weeks. But we probably shouldn't be just grabbing the latest version of gcloud on every build anyway.